I did a lot of things.
I got better textures for the Everfree Forest.  There are changelings in
place of the timberwolves. We have our first attack, which is fire. It's
unfinished, but appears when you left click.  We have simple enemies,
which are buggy as taurus, due to the fact that they create unwanted
rotations. We should get a better system later. We have a house sprite.
We have collisions for enemies and the player in the Everfree Forest.
There may have been more added, but that is the majority. There are no
bugs, aside from the changelings in this version.